requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
cryptography
